[modbus] Move some helpers to shared location
