+++ esphome/components/modbus/modbus_helpers.h
@@ -1,0 +1,121 @@
+#pragma once
+
+#include <string>
+#include <vector>
+#include <cmath>
+
+#include "esphome/core/helpers.h"
+#include "esphome/components/modbus/modbus_definitions.h"
+
+namespace esphome {
+namespace modbus {
+namespace helpers {
+
+enum class SensorValueType : uint8_t {
+  RAW = 0x00,     // variable length
+  U_WORD = 0x1,   // 1 Register unsigned
+  U_DWORD = 0x2,  // 2 Registers unsigned
+  S_WORD = 0x3,   // 1 Register signed
+  S_DWORD = 0x4,  // 2 Registers signed
+  BIT = 0x5,
+  U_DWORD_R = 0x6,  // 2 Registers unsigned
+  S_DWORD_R = 0x7,  // 2 Registers unsigned
+  U_QWORD = 0x8,
+  S_QWORD = 0x9,
+  U_QWORD_R = 0xA,
+  S_QWORD_R = 0xB,
+  FP32 = 0xC,
+  FP32_R = 0xD
+};
+
+inline bool value_type_is_float(SensorValueType v) {
+  return v == SensorValueType::FP32 || v == SensorValueType::FP32_R;
+}
+
+inline ModbusFunctionCode modbus_register_read_function(ModbusRegisterType reg_type) {
+  switch (reg_type) {
+    case ModbusRegisterType::COIL:
+      return ModbusFunctionCode::READ_COILS;
+      break;
+    case ModbusRegisterType::DISCRETE_INPUT:
+      return ModbusFunctionCode::READ_DISCRETE_INPUTS;
+      break;
+    case ModbusRegisterType::HOLDING:
+      return ModbusFunctionCode::READ_HOLDING_REGISTERS;
+      break;
+    case ModbusRegisterType::READ:
+      return ModbusFunctionCode::READ_INPUT_REGISTERS;
+      break;
+    default:
+      return ModbusFunctionCode::CUSTOM;
+      break;
+  }
+}
+
+inline ModbusFunctionCode modbus_register_write_function(ModbusRegisterType reg_type) {
+  switch (reg_type) {
+    case ModbusRegisterType::COIL:
+      return ModbusFunctionCode::WRITE_SINGLE_COIL;
+      break;
+    case ModbusRegisterType::DISCRETE_INPUT:
+      return ModbusFunctionCode::CUSTOM;
+      break;
+    case ModbusRegisterType::HOLDING:
+      return ModbusFunctionCode::READ_WRITE_MULTIPLE_REGISTERS;
+      break;
+    case ModbusRegisterType::READ:
+    default:
+      return ModbusFunctionCode::CUSTOM;
+      break;
+  }
+}
+
+inline uint8_t c_to_hex(char c) { return (c >= 'A') ? (c >= 'a') ? (c - 'a' + 10) : (c - 'A' + 10) : (c - '0'); }
+
+/** Get a byte from a hex string
+ *  hex_byte_from_str("1122",1) returns uint_8 value 0x22 == 34
+ *  hex_byte_from_str("1122",0) returns 0x11
+ * @param value string containing hex encoding
+ * @param position  offset in bytes. Because each byte is encoded in 2 hex digits the position of the original byte in
+ * the hex string is byte_pos * 2
+ * @return byte value
+ */
+inline uint8_t byte_from_hex_str(const std::string &value, uint8_t pos) {
+  if (value.length() < pos * 2 + 1)
+    return 0;
+  return (c_to_hex(value[pos * 2]) << 4) | c_to_hex(value[pos * 2 + 1]);
+}
+
+/** Get a word from a hex string
+ * @param value string containing hex encoding
+ * @param position  offset in bytes. Because each byte is encoded in 2 hex digits the position of the original byte in
+ * the hex string is byte_pos * 2
+ * @return word value
+ */
+inline uint16_t word_from_hex_str(const std::string &value, uint8_t pos) {
+  return byte_from_hex_str(value, pos) << 8 | byte_from_hex_str(value, pos + 1);
+}
+
+/** Get a dword from a hex string
+ * @param value string containing hex encoding
+ * @param position  offset in bytes. Because each byte is encoded in 2 hex digits the position of the original byte in
+ * the hex string is byte_pos * 2
+ * @return dword value
+ */
+inline uint32_t dword_from_hex_str(const std::string &value, uint8_t pos) {
+  return word_from_hex_str(value, pos) << 16 | word_from_hex_str(value, pos + 2);
+}
+
+/** Get a qword from a hex string
+ * @param value string containing hex encoding
+ * @param position  offset in bytes. Because each byte is encoded in 2 hex digits the position of the original byte in
+ * the hex string is byte_pos * 2
+ * @return qword value
+ */
+inline uint64_t qword_from_hex_str(const std::string &value, uint8_t pos) {
+  return static_cast<uint64_t>(dword_from_hex_str(value, pos)) << 32 | dword_from_hex_str(value, pos + 4);
+}
+
+}  // namespace helpers
+}  // namespace modbus
+}  // namespace esphome
