+++ esphome/components/modbus_controller/select/__init__.py
@@ -1,15 +1,10 @@
 import esphome.codegen as cg
 from esphome.components import select
+from esphome.components.modbus.helpers import SENSOR_VALUE_TYPE, TYPE_REGISTER_MAP
 import esphome.config_validation as cv
 from esphome.const import CONF_ADDRESS, CONF_ID, CONF_LAMBDA, CONF_OPTIMISTIC
 
-from .. import (
-    SENSOR_VALUE_TYPE,
-    TYPE_REGISTER_MAP,
-    ModbusController,
-    SensorItem,
-    modbus_controller_ns,
-)
+from .. import ModbusController, SensorItem, modbus_controller_ns
 from ..const import (
     CONF_FORCE_NEW_RANGE,
     CONF_MODBUS_CONTROLLER_ID,
