+++ esphome/components/modbus_controller/sensor/__init__.py
@@ -1,11 +1,10 @@
 import esphome.codegen as cg
 from esphome.components import sensor
+from esphome.components.modbus.helpers import MODBUS_REGISTER_TYPE, SENSOR_VALUE_TYPE
 import esphome.config_validation as cv
 from esphome.const import CONF_ADDRESS, CONF_ID
 
 from .. import (
-    MODBUS_REGISTER_TYPE,
-    SENSOR_VALUE_TYPE,
     ModbusItemBaseSchema,
     SensorItem,
     add_modbus_base_properties,
